requirements.txt
amqp==5.3.1
asgiref==3.10.0
attrs==25.4.0
billiard==4.2.4
celery==5.3.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
diff-match-patch==20241021
dj-database-url==2.1.0
Django==5.2.9
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==4.3.14
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
et_xmlfile==2.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
pillow==11.0.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rest-framework-simplejwt==0.0.2
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
whitenoise==6.6.0
XlsxWriter==3.1.9
